+++ aiohomematic/central/coordinators/hub.py
@@ -43,6 +43,7 @@ from aiohomematic.interfaces import (
 )
 from aiohomematic.model.hub import Hub, InstallModeDpType, MetricsDpType, ProgramDpType
 from aiohomematic.property_decorators import DelegatedProperty
+from aiohomematic.type_aliases import UnsubscribeCallback
 
 _LOGGER: Final = logging.getLogger(__name__)
 
@@ -58,6 +59,7 @@ class HubCoordinator(HubDataFetcherProtocol, HubDataPointManagerProtocol):
         "_program_data_points",
         "_state_path_to_name",
         "_sysvar_data_points",
+        "_sysvar_unsubscribes",
     )
 
     def __init__(
@@ -102,6 +104,8 @@ class HubCoordinator(HubDataFetcherProtocol, HubDataPointManagerProtocol):
         # {program_name, program_button}
         self._program_data_points: Final[dict[str, ProgramDpType]] = {}
         self._state_path_to_name: Final[dict[str, str]] = {}
+        # Unsubscribe callbacks for sysvar event subscriptions
+        self._sysvar_unsubscribes: Final[list[UnsubscribeCallback]] = []
 
         # Create Hub with protocol interfaces
         self._hub: Final = Hub(
@@ -183,9 +187,21 @@ class HubCoordinator(HubDataFetcherProtocol, HubDataPointManagerProtocol):
                 if event.state_path == sysvar_data_point.state_path:
                     await sysvar_data_point.event(value=event.value, received_at=event.received_at)
 
-            self._event_bus_provider.event_bus.subscribe(
+            unsubscribe = self._event_bus_provider.event_bus.subscribe(
                 event_type=SysvarStateChangedEvent, event_key=sysvar_data_point.state_path, handler=event_handler
             )
+            self._sysvar_unsubscribes.append(unsubscribe)
+
+    def clear(self) -> None:
+        """
+        Clear all sysvar event subscriptions.
+
+        Call this method when stopping the central unit to prevent leaked subscriptions.
+        """
+        for unsubscribe in self._sysvar_unsubscribes:
+            unsubscribe()
+        self._sysvar_unsubscribes.clear()
+        _LOGGER.debug("CLEAR: Cleared %s sysvar event subscriptions", self._central_info.name)
 
     def create_install_mode_dps(self) -> Mapping[Interface, InstallModeDpType]:
         """
@@ -450,7 +466,7 @@ class HubCoordinator(HubDataFetcherProtocol, HubDataPointManagerProtocol):
             self._sysvar_data_points.pop(vid, None)
             self._state_path_to_name.pop(sysvar_dp.state_path, None)
 
-            # Event subscription will be automatically cleaned up when sysvar_dp is deleted
+            # Note: Event subscriptions are cleaned up via clear() when central stops
 
             _LOGGER.debug(
                 "REMOVE_SYSVAR_DATA_POINT: Removed sysvar %s from %s",